+++ src/kaocha/stacktrace.clj
@@ -7,11 +7,16 @@
                                      "clojure.lang."
                                      "clojure.core"
                                      "clojure.main"
-                                     "orchestra."
                                      "kaocha.monkey_patch"])
 
 (defn elide-element? [e]
   (some #(str/starts-with? (.getClassName ^StackTraceElement e) %) *stacktrace-filters*))
+
+(def ^:dynamic *stacktrace-stop-list* ["kaocha.ns"
+                                       "lambdaisland.tools.namespace.reload"])
+
+(defn sentinel-element? [e]
+  (some #(str/starts-with? (.getClassName ^StackTraceElement e) %) *stacktrace-stop-list*))
 
 (defn print-stack-trace
   "Prints a Clojure-oriented stack trace of tr, a Throwable.
@@ -35,16 +40,18 @@
          (let [n (cond-> n n dec)]
            (if (= 0 n)
              (println "    ... and " (count st) "more")
-             (if (elide-element? e)
-               (do
-                 (when (not eliding?)
-                   (println "    ..."))
-                 (recur st true n))
-               (do
-                 (print "    ")
-                 (st/print-trace-element e)
-                 (newline)
-                 (recur st false n))))))))))
+             (if (sentinel-element? e)
+               (println "(Rest of stacktrace elided)")
+               (if (elide-element? e)
+                 (do
+                   (when (not eliding?)
+                     (println "    ..."))
+                   (recur st true n))
+                 (do
+                   (print "    ")
+                   (st/print-trace-element e)
+                   (newline)
+                   (recur st false n)))))))))))
 
 (defn print-cause-trace
   "Like print-stack-trace but prints chained exceptions (causes)."