+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+## Added
+
+## Fixed
+
+## Changed
+
 # 1.87.1366 (2023-09-27 / f514905)
 
 ## Added