+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Unreleased
 
-## Added
-
 ## Fixed
 
 - Track all results when `:once` fixtures call `f` multiple times.
   (thanks [@NoahTheDuke](https://github.com/NoahTheDuke))
 
 ## Changed
+
+- Version bumps
 
 # 1.87.1366 (2023-09-27 / f514905)