+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+## Added
+
+## Fixed
+
+## Changed
+
 # 1.88.1376 (2024-03-25 / 8308e16)
 
 ## Fixed